new loplugin:droplong
The types 'long' and 'unsigned long' are different sizes on different
platforms, making them wholy unsuitable for portable code.
And when I mean different sizes, I mean 64bit Linux and 64bit Windows
have different bit sizes.

Reviewed-on: https://gerrit.libreoffice.org/41130
Tested-by: Jenkins <ci@libreoffice.org>

+++ codemaker/source/javamaker/javaoptions.cxx
@@ -208,7 +208,7 @@ bool JavaOptions::initOptions(int ac, char* av[], bool bCmdFile)
 
                     ret = initOptions(rargc, rargv, bCmdFile);
 
-                    for (long j=0; j < rargc; j++)
+                    for (int j=0; j < rargc; j++)
                     {
                         free(rargv[j]);
                     }
